+++ source/Makefile.Objects
@@ -746,9 +746,9 @@ OBJS_SRCPW=h_ewald_pw.o\
     setup_pwwfc.o\
     update_cell_pw.o\
     dftu_base.o\
-    dftu_output.o\
-    dftu_tools_pw.o\
-    dftu_cal_occ_pw.o\
+    dftu_base_io.o\
+    dftu_base_occ.o\
+    dftu_base_tools.o\
     setup_dftu_pw.o\
     deltaspin_pw.o\
     force_pw.o\
