+++ source/source_lcao/setup_dftu_lcao.cpp
@@ -3,7 +3,7 @@
 #include "source_lcao/module_dftu/dftu_lcao_occ.h"
 #include "source_lcao/module_dftu/dftu_lcao_energy.h"
 #include "source_lcao/module_dftu/dftu_yukawa.h"
-#include "source_pw/module_pwdft/dftu_output.h" // mohan add 2025-11-08
+#include "source_pw/module_pwdft/dftu_base_io.h" // mohan add 2025-11-08
 #include "source_estate/module_dm/density_matrix.h"
 #include "source_lcao/hamilt_lcao.h"
 
@@ -73,7 +73,7 @@ void finish_dftu_lcao(const int iter,
         }
         DFTU_LCAO::cal_energy_correction(*dftu_ptr, ucell);
     }
-    dftu_io::output(*dftu_ptr, ucell, out_chg, global_out_dir, nspin, npol);
+    DFTU_BASE::output(*dftu_ptr, ucell, out_chg, global_out_dir, nspin, npol);
     
     /// use the converged occupation matrix for next MD/Relax SCF calculation
     if (conv_esolver)
